Crate dock_testnet_runtime

Dock testnet runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

blob
Generic immutable single-owner storage.

did

opaque
Opaque types. These are used by the CLI to instantiate machinery that don't need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

revoke

Structs

AvailableBlockRatio

BlockHashCount

ExistentialDeposit

GenesisConfig

MaxBlobSize

MaximumBlockLength

MaximumBlockWeight
We allow for 2 seconds of compute with a 6 second average block time.

MaximumExtrinsicWeight
Assume 10% of weight for average on_initialize calls.

MinimumPeriod

ModuleToIndex
Provides an implementation of ModuleToIndex to map a module to its index in the runtime.

Runtime

RuntimeApi

RuntimeApiImpl
Implements all runtime apis for the client side.

TransactionByteFee

Version

Enums

Call

Event

Origin

StateChange
Any state change that needs to be signed is first wrapped in this enum and then its serialized. This is done to prevent make it unambiguous which command was intended as the SCALE codec's not self describing. Never change the order of variants in this enum

Constants

VERSION
This runtime version.

WASM_BINARY

Functions

native_version
The version information used to identify this runtime when compiled natively.

Type Definitions

Aura

AuraConfig

Balances

BalancesConfig

BlobStore

DIDModule

Grandpa

GrandpaConfig

RandomnessCollectiveFlip

Revoke

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionPayment