Module dock_testnet_runtime::opaque

Opaque types. These are used by the CLI to instantiate machinery that don't need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core datastructures.

Structs

SessionKeys

Type Definitions

Block
Opaque block type.