Module dock_testnet_runtime::revoke

Structs

Module

Registry
Metadata about a revocation scope.

RemoveRegistry
Command to remove an entire registy. Removes all revocations in the registry as well as registry metadata.

Revoke
Command to create a set of revocations withing a registry. Creation of revocations is idempotent; creating a revocation that already exists is allowed, but has no effect.

UnRevoke
Command to remove a set of revocations within a registry. Removal of revocations is idempotent; removing a revocation that doesn't exists is allowed, but has no effect.

Enums

Call
Dispatchable calls.

Policy
Authorization logic for a registry.

RevErr
Revocation Error

Traits

Trait

Type Definitions

PAuth
Proof of authorization to modify a registry.

RegistryId
Points to an on-chain revocation registry.

RevokeId
Points to a revocation which may or may not exist in a registry.